% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispit_pvalue.R
\name{multisplit.pval}
\alias{multisplit.pval}
\title{Mediation analysis using partial sum statistic and sample splitting}
\usage{
multisplit.pval(pvalue, min.gamma = 0.05)
}
\arguments{
\item{pvalue}{XXX}

\item{min.gamma}{XXXX}
}
\value{
XXXXXX
}
\description{
aggregate p-value from multi-split result
}
\details{
write XXX
}
\seealso{
\code{\link{PS5_multi_split}}
}
\author{
Hung-Ching Chang
}
