% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PS5_mediation.R
\name{PS5.mediation}
\alias{PS5.mediation}
\title{Mediation analysis using partial sum statistic and sample splitting}
\usage{
PS5.mediation(
  M = mediators,
  X = exposure,
  Y = outcome,
  C = conf,
  n.draw = 10000,
  dim.reduction = NA,
  seed = 1004,
  ga = 2,
  Y.family = "gaussian",
  M.family = "gaussian",
  pre.selection = "MCP",
  lambda.selection = "BIC"
)
}
\arguments{
\item{M}{# n-by-p matrix}

\item{X}{# n      vector}

\item{Y}{# n      vector}

\item{C}{# n-by-l matrix, not including intercept}

\item{n.draw}{# no. of parametric bootstrapping sample}

\item{dim.reduction, }{# dimension reduction}

\item{ga}{# gamma}

\item{lambda.selection}{}
}
\value{
A \code{list} of containing mediation testing result.
\itemize{
    \item{global.test: }{}
    \item{global.me: }{}
    \item{global.me.prop: }{}
    \item{estimation: }{}
}
}
\description{
XXXXX
}
\details{
write XXX
}
\seealso{
\code{\link{sample.splitting.dim.reduction}, \link{PS5_multi_split}}
}
\author{
Hung-Ching Chang
}
