% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PS5_multi_split.R
\name{PS5.multi.split}
\alias{PS5.multi.split}
\title{Mediation analysis using partial sum statistic and sample splitting}
\usage{
PS5.multi.split(
  M = mediators,
  X = exposure,
  Y = outcome,
  C = conf,
  n.draw = 10000,
  dim.reduction = NA,
  seed = 1004,
  Y.family = "gaussian",
  M.family = "gaussian",
  pre.selection = "MCP",
  multi.num = 100,
  cores = detectCores() - 1,
  ga = 2
)
}
\arguments{
\item{M}{# n-by-p matrix}

\item{X}{# n      vector}

\item{Y}{# n      vector}

\item{C}{# n-by-l matrix, not including intercept}

\item{n.draw}{# no. of parametric bootstrapping sample}

\item{dim.reduction, }{# dimension reduction}

\item{cores}{}

\item{ga}{# gamma}
}
\value{
XXXXXX
}
\description{
aggregate p-value from multi-split result
}
\details{
write XXX
}
\seealso{
\code{\link{PS5_mediation}}
}
\author{
Hung-Ching Chang
}
