% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PS5multisplit.R
\name{PS5.multisplit}
\alias{PS5.multisplit}
\title{High-dimensional mediation analysis by multiple sample splitting}
\usage{
PS5.multisplit(
  M = mediators,
  X = exposure,
  Y = outcome,
  C = conf,
  n.draw = 10000,
  dim.reduction = TRUE,
  seed = 1004,
  ga = 2,
  Y.family = "gaussian",
  M.family = "gaussian",
  penalty = "MCP",
  multi.num = 100,
  cores = 1
)
}
\arguments{
\item{M}{a matrix of high-dimensional mediators. Row is samples, and column is variables.}

\item{X}{a vector of exposure.}

\item{Y}{a vector of outcome.}

\item{C}{a matrix of confounding variables}

\item{n.draw}{number of draws for parametric bootstrap. Default is 10,000.}

\item{dim.reduction}{a logical value. If 'TRUE' (default) a sample splitting strategy with penalized regression regression is used for dimension reduction; if 'FALSE' dimension reduction procedure will be skipped..}

\item{seed}{a random seed. Default is 1004.}

\item{ga}{value of gamma parameter. Default is 2.}

\item{Y.family}{either 'gaussian' (default) or 'binomial', depending on the data type of outcome (Y)}

\item{M.family}{either 'gaussian' (default) or 'binomial', depending on the data type of mediator (M)}

\item{penalty}{the penalty method. Either 'MCP' (the default), 'SCAD', or 'lasso'.}

\item{multi.num}{number of multi sample split. Default is 100.}

\item{cores}{number of cores in parallel computing. Default is 1.}
}
\value{
A \code{list} containing mediation testing result.
\itemize{
    \item{global.test: }{statistical significance of global mediation effect}
    \item{global.me: }{estimated global mediation effect}
    \item{global.me.pct: }{estimated global mediation percentage}
    \item{mediation.contri: }{a data.frame of individual mediation contribution}
    \item{detail.global.test: }{a vector containing global.test from each sample split}
    \item{detail.global.me: }{a vector containing global.me from each sample split}
    \item{detail.global.me.pct: }{a vector containing global.me.pct from each sample split}
    \item{detail.mediation.contri: }{ A \code{list} containing mediation.contri from each sample split}
}
}
\description{
\code{PS5.multisplit} enhances the robustness through multi sample split for estimating and testing global mediation effect and individual mediation contribution.
}
\examples{
XXXXXXXXXX

}
\author{
Hung-Ching Chang
}
