% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.splitting.dim.reduction.R
\name{sample.splitting.dim.reduction}
\alias{sample.splitting.dim.reduction}
\title{Mediation analysis using partial sum statistic and sample splitting}
\usage{
sample.splitting.dim.reduction(
  M,
  X,
  Y,
  C,
  Y.family = "gaussian",
  penalty.type = "MCP",
  lambda.selection = "BIC"
)
}
\arguments{
\item{lambda.selection}{}
}
\value{
XXXXXX
}
\description{
aggregate p-value from multi-split result
}
\details{
write XXX
}
\seealso{
\code{\link{PS5_mediation}}
}
\author{
Hung-Ching Chang
}
