% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PS5.R
\name{PS5}
\alias{PS5}
\title{High-dimensional mediation analysis by single sample splitting}
\usage{
PS5(
  M = mediators,
  X = exposure,
  Y = outcome,
  C = conf,
  n.draw = 10000,
  dim.reduction = TRUE,
  seed = 1004,
  ga = 2,
  Y.family = "gaussian",
  M.family = "gaussian",
  penalty = "MCP"
)
}
\arguments{
\item{M}{a matrix of high-dimensional mediators. Row is samples, and column is variables.}

\item{X}{a vector of exposure.}

\item{Y}{a vector of outcome.}

\item{C}{a matrix of confounding variables}

\item{n.draw}{number of draws for parametric bootstrap}

\item{dim.reduction}{a logical value. If 'TRUE' a sample splitting strategy with penalized regression regression is used for dimension reduction; if 'FALSE' dimension reduction procedure will be skipped. Default is 'TRUE'.}

\item{seed}{a random seed}

\item{ga}{value of gamma parameter. Default is 2.}

\item{Y.family}{either 'gaussian' (default) or 'binomial', depending on the data type of outcome (Y)}

\item{M.family}{either 'gaussian' (default) or 'binomial', depending on the data type of mediator (M)}

\item{penalty}{the penalty method. Either 'MCP' (the default), 'SCAD', or 'lasso'.}
}
\value{
A \code{list} of containing mediation testing result.
\itemize{
    \item{global.test: }{}
    \item{global.me: }{}
    \item{global.me.prop: }{}
    \item{estimation: }{}
}
}
\description{
\code{PS5} is used to estimate and test global mediation effect and individual mediation contribution for high-dimensional causal mediation analysis.
}
\details{
XXXXXXXXXXXXXXXXXXXXXX
}
\author{
Hung-Ching Chang
}
